% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_url.R
\name{create_url}
\alias{create_url}
\title{Create the URL for the ZAMG data hub API}
\usage{
create_url(type, mode, resource_id)
}
\arguments{
\item{type}{Data type, as available in \code{\link{datasets_zamg}}.}

\item{mode}{Data mode, as available in \code{\link{datasets_zamg}}.}

\item{resource_id}{The resource_id, as available in \code{\link{datasets_zamg}} .}
}
\value{
The URL as a character string.
}
\description{
Build the URL to access the API by specifying each part. The needed
information is available as internal dataset \code{\link{datasets_zamg}} with
column names the same as function parameters.
}
\examples{
# The first part of the URL as in the example on
# https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html

create_url(type = "station",
           mode = "historical",
           resource_id = "klima-v1-10min")

}
