% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_url.R
\name{data_from_url}
\alias{data_from_url}
\title{Download and parse data from URL}
\usage{
data_from_url(u)
}
\arguments{
\item{u}{A request, as created by \code{\link{create_url}}.}
}
\value{
A list with parameter info, station location, and data (as tibbles).
}
\description{
Transforms the JSON output of a request created by \code{\link{create_url}}
into a list that holds almost all the returned information (parameter info,
station location, the actual data).
}
\examples{
\dontrun{

# The same URL as in the example on
# https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html

u <- create_url(resource_id = "klima-v1-10min",
                parameters = c("TL", "RR", "RRM"),
                date_start = "2020-12-24T08:00",
                date_end = "2020-12-24T09:00",
                station_ids = "5904")

data_from_url(u = u)

}

}
