% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_zamg.R
\docType{data}
\name{datasets_zamg}
\alias{datasets_zamg}
\title{Available datasets in the ZAMG data hub}
\format{
A data frame with 63 rows and 4 variables:
\describe{
  \item{resource_id}{id, name of the resource}
  \item{type}{station, grid, or timeseries}
  \item{mode}{historical, current, or forecast}
  \item{response_formats}{geojson, netcdf, csv}
}
}
\source{
\url{https://dataset.api.hub.zamg.ac.at/v1/docs/index.html}
}
\usage{
datasets_zamg
}
\description{
Information needed to build and check the URLs to access the API.
}
\keyword{datasets}
