% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Download metadata from ZAMG API}
\usage{
get_metadata(type = "station", mode = "historical", resource_id)
}
\arguments{
\item{type}{Data type, currently, pkg is only tested with "station" (default).}

\item{mode}{Data mode, currently, pkg is only tested with "historical" (default).}

\item{resource_id}{The resource_id.}
}
\value{
A list with two elements, the first a tibble of parameter information,
  the second with station information.
}
\description{
Mostly for internal use, since the output is already available as
internal data sets: \code{\link{datasets_zamg}} and  \code{\link{metadata_zamg}}.
}
\details{
A list of available data sets (with resource_id) can also be retrieved
from \url{http://dataset.api.hub.zamg.ac.at/v1/datasets/}.
}
\examples{

get_metadata(resource_id = "klima-v1-10min")


}
