% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_funs.R
\name{get_metadata_station}
\alias{get_metadata_station}
\alias{get_metadata_grid}
\title{Download metadata from ZAMG API}
\usage{
get_metadata_station(u)

get_metadata_grid(u)
}
\arguments{
\item{u}{URL from \code{\link{create_url}}.}
}
\value{
A list with two elements, the first a tibble of parameter information,
  the second with station/grid information.
}
\description{
For internal use, since the output is already available as
internal data sets: \code{\link{datasets_zamg}} and  \code{\link{metadata_station}}.
}
\details{
A list of available data sets (with resource_id) can also be retrieved
from \url{http://dataset.api.hub.zamg.ac.at/v1/datasets/}.
}
