% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_zamg.R
\docType{data}
\name{metadata_zamg}
\alias{metadata_zamg}
\title{Metadata of stations and parameters}
\format{
The first item (named parameters) holds a tibble of all available parameters
for this resource_id.

\describe{
  \item{name}{Parameter name, used in building the query}
  \item{long_name}{German long name}
  \item{desc}{German description of the paramater}
  \item{unit}{self-explaining}
}

The second item (named stations) holds a tibble of all available stations
for this resource_id.

\describe{
  \item{type}{COMBINED or INDIVIDUAL series}
  \item{id}{Station id, used in the query}
  \item{group_id}{Identifies groups of stations that belong together (e.g., relocations)}
  \item{name}{Station name}
  \item{state}{Austrian state (Bundesland)}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  \item{altitutde}{Elevation a.s.l.}
  \item{valid_from}{Date start of observations}
  \item{valid_to}{Date end of observations (? 2100-12-01 might imply up to now)}
}
}
\source{
\url{https://dataset.api.hub.zamg.ac.at/v1/docs/index.html}
}
\usage{
metadata_zamg
}
\description{
A named list with resource_ids corresponding to station time series (klima*).
Each list item contains again a list of length two.
}
\keyword{datasets}
