% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_grid.R
\name{query_grid}
\alias{query_grid}
\title{Add query for grid data to the base URL}
\usage{
query_grid(u, parameters, date_start, date_end, bbox, output = "netcdf")
}
\arguments{
\item{u}{URL from \code{\link{create_url}}.}

\item{parameters}{Parameters (character), might have different name depending
on resource_id, see also \code{\link{metadata_grid}}.}

\item{date_start}{The start date (ISO8601) to download the time series.}

\item{date_end}{The end date (ISO8601) to download the time series.}

\item{bbox}{Bounding box, numeric vector of length 4 with special order:
(ymin, xmin, ymax, xmax). See also \code{\link{metadata_grid}} for extent.}

\item{output}{Output format: netcdf (default) or geojson.}
}
\value{
The request (URL + query) as a character string.
}
\description{
Specify each part of the query and append to the URL, which is created by
\code{\link{create_url}}. See \code{\link{metadata_grid}} for help to derive
the needed parameters. For how to specify dates, see
\url{https://en.wikipedia.org/wiki/ISO_8601}.
}
\examples{

library(magrittr)

create_url(type = "grid", mode = "historical", resource_id = "spartacus-v1-1d-1km") \%>\%
    query_grid(c("RR", "Tn"), "2020-09-01", "2020-09-05", c(47.3, 13.38, 48.2, 14.4))

}
