% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_station.R
\name{query_station}
\alias{query_station}
\title{Add query for station data to the base URL}
\usage{
query_station(u, parameters, date_start, date_end, station_ids)
}
\arguments{
\item{u}{URL from \code{\link{create_url}}.}

\item{parameters}{Parameters (character), might have different name depending
on resource_id, see also \code{\link{metadata_station}}.}

\item{date_start}{The start date (ISO8601) to download the time series.}

\item{date_end}{The end date (ISO8601) to download the time series.}

\item{station_ids}{Station identifiers (character), see also
\code{\link{metadata_station}}.}
}
\value{
The request (URL + query) as a character string.
}
\description{
Specify each part of the query and append to the URL, which is created by
\code{\link{create_url}}. See \code{\link{metadata_station}} for help to
derive the needed parameters. For how to specify dates, see
\url{https://en.wikipedia.org/wiki/ISO_8601}.
}
\examples{

library(magrittr)
# The same URL as in the example on
# https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html

create_url(type = "station", mode = "historical", resource_id = "klima-v1-10min") \%>\%
    query_station(parameters = c("TL", "RR", "RRM"),
                  date_start = "2020-12-24T08:00",
                  date_end = "2020-12-24T09:00",
                  station_ids = "5904")

}
