% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_series.R
\name{single_series}
\alias{single_series}
\title{Download a single time series of one parameter}
\usage{
single_series(
  resource_id,
  parameter,
  date_start,
  date_end,
  station_id,
  type = "station",
  mode = "historical"
)
}
\arguments{
\item{resource_id}{The resource_id as available in \code{\link{datasets_zamg}}.}

\item{parameter}{Parameter (character, length 1), might have different name depending on resource_id,
see also \code{\link{metadata_zamg}}.}

\item{date_start}{The start date (ISO8601) to download the time series.}

\item{date_end}{The end date (ISO8601) to download the time series.}

\item{station_id}{Station identifier (character, length 1), see also \code{\link{metadata_zamg}}.}

\item{type}{Data type, currently, pkg is only tested with "station" (default).}

\item{mode}{Data mode, currently, pkg is only tested with "historical" (default).}
}
\value{
A tibble with columns datetime and value.
}
\description{
A streamlined combination of \code{\link{create_url}} and
\code{\link{data_from_url}} that only allows one station
and one parameter to be downloaded. Only returns the time
series without metadata.
}
\examples{
single_series(resource_id = "klima-v1-10min",
              parameter = "RR",
              date_start = "2020-12-24T08:00",
              date_end = "2020-12-24T09:00",
              station_id = "5904")

}
