% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\name{station_data}
\alias{station_data}
\title{Download and parse station data from URL}
\usage{
station_data(request)
}
\arguments{
\item{request}{A request, as created by \code{\link{query_station}}.}
}
\value{
A list with parameter info, station location, and data (as tibbles).
}
\description{
Transforms the JSON output of a request created by \code{\link{query_station}}
into a list that holds almost all the returned information (parameter info,
station location, the actual data).
}
\examples{

library(magrittr)
# The same request as in the example on
# https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html

create_url(type = "station", mode = "historical", resource_id = "klima-v1-10min") \%>\%
    query_station(parameters = c("TL", "RR", "RRM"),
                  date_start = "2020-12-24T08:00",
                  date_end = "2020-12-24T09:00",
                  station_ids = "5904") \%>\%
    station_data()


}
